First upload
This is the first version of the code. The file "Tidying_Data.py" takes downloaded CSV files with Svensson parameters and calculates the time series of interest rates and the term structure of interest rates based on the parameters. 

The file "Visualizing_Time_Series_of_Interest_Rates.py" visualizes the time series. 

There is no file for visualizing the term structure of interest rates yet.