Added Option to show values on second y-axis
I got a request from the Department of Finance to compare the 5-year yield spread to the development of the VDAX. The VDAX (or any time series of variables) can now be shown on the second y-axis on the right. Detailed instructions on implementing this within the code can be taken from the README file.